rest_extract/requirements.txt
#rest_extract
pdfminer.six
flask
pyyaml
genson
requests
